Fixed bug where SCUserInfo.avatarURL contained a NSString instead of NSURL. (how bad is the type safety…) Added ability to add user to database. Added coverURL to database. Small changes relating to debugging and code cleanliness/consistency

+++ Matar/SCUserInfo.h
@@ -10,7 +10,7 @@
 
 @interface SCUserInfo : NSObject
 
-@property int ID;
+@property NSInteger ID;
 @property (strong) NSString *userName;
 @property (strong) NSURL *avatarURL;
